requirements.txt
flask==2.3.3
pytesseract==0.3.10
pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
transformers==4.32.1
google-cloud-translate==3.12.0
opencv-python==4.8.0.76
scikit-image==0.21.0
langdetect==1.0.9
pdfplumber==0.9.0
deepl==1.15.0
spacy==3.6.1
matplotlib==3.7.2
waitress==2.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
tqdm==4.66.1